requirements.txt
numpy==1.26.4
pillow==10.2.0
tensorflow==2.15.0
tqdm==4.66.2